requirements.txt

Django==4.0.5
django-paranoid==1.0
djangorestframework==3.13.1